Use uint32_t fpr reservedSize

+++ src/app/CommandHandlerImpl.cpp
@@ -74,10 +74,10 @@ CHIP_ERROR CommandHandlerImpl::AllocateBuffer()
         // On some platforms we can get more available length in the packet than what we requested.
         // It is vital that we only use up to commandBufferMaxSize for the entire packet and
         // nothing more.
-        uint16_t reservedSize = 0;
+        uint32_t reservedSize = 0;
         if (commandPacket->AvailableDataLength() > commandBufferMaxSize)
         {
-            reservedSize = static_cast<uint16_t>(commandPacket->AvailableDataLength() - commandBufferMaxSize);
+            reservedSize = static_cast<uint32_t>(commandPacket->AvailableDataLength() - commandBufferMaxSize);
         }
 
         mCommandMessageWriter.Init(std::move(commandPacket));
@@ -91,7 +91,7 @@ CHIP_ERROR CommandHandlerImpl::AllocateBuffer()
         // MIC might require up to kMaxTagLen. We need to make sure to reserve this space
         // at the end of the buffer.
         ReturnErrorOnFailure(mInvokeResponseBuilder.GetWriter()->ReserveBuffer(
-            static_cast<uint32_t>(reservedSize + MessagePacketBuffer::kMaxFooterSize)));
+            reservedSize + MessagePacketBuffer::kMaxFooterSize));
 
         // Sending an InvokeResponse to an InvokeResponse is going to be removed from the spec soon.
         // It was never implemented in the SDK, and there are no command responses that expect a
